[FEAT] Apply Move Activity Func
